requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
httpx
python-dotenv
pdfplumber
reportlab
jinja2
